GUAC-1132: Rename DirectoryObject to ModeledDirectoryObject.
