+++ extensions/guacamole-auth-jdbc/modules/guacamole-auth-jdbc-base/src/main/java/org/glyptodon/guacamole/auth/jdbc/base/ModeledDirectoryObjectMapper.java
@@ -37,7 +37,7 @@ import org.apache.ibatis.annotations.Param;
  *     The type of object contained within the directory whose objects are
  *     mapped by this mapper.
  */
-public interface DirectoryObjectMapper<ModelType> {
+public interface ModeledDirectoryObjectMapper<ModelType> {
 
     /**
      * Selects the identifiers of all objects, regardless of whether they
