+++ extensions/guacamole-auth-jdbc/modules/guacamole-auth-jdbc-base/src/main/java/org/glyptodon/guacamole/auth/jdbc/base/ModeledGroupedDirectoryObject.java
@@ -33,8 +33,8 @@ import org.glyptodon.guacamole.auth.jdbc.connectiongroup.RootConnectionGroup;
  * @param <ModelType>
  *     The type of model object that corresponds to this object.
  */
-public abstract class GroupedDirectoryObject<ModelType extends GroupedObjectModel>
-    extends DirectoryObject<ModelType> {
+public abstract class ModeledGroupedDirectoryObject<ModelType extends GroupedObjectModel>
+    extends ModeledDirectoryObject<ModelType> {
 
     /**
      * Returns the identifier of the parent connection group, which cannot be
